# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate a user and return a JWT token.",
                "operationId": "c5dc3a8853351d3147a0af6b82d08f19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login, returns JWT token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email or password."
                    }
                }
            }
        },
...
        "/api/v1/payment": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "List payments with pagination",
                "operationId": "7c04a5c5f51834b6621e7aa860935026",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of payments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "01971ef6-df01-723e-85bf-be4ca4b95685"
                                                    },
                                                    "name_client": {
                                                        "type": "string",
                                                        "example": "Cliente Teste"
                                                    },
                                                    "cpf": {
                                                        "type": "string",
                                                        "example": "88811188888"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "amount": {
                                                        "type": "string",
                                                        "example": "10.55"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "method_payment": {
                                                        "type": "string",
                                                        "example": "boleto"
                                                    },
                                                    "paid_at": {
                                                        "type": "string",
                                                        "example": "2025-05-29 23:13:04"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/v1/payment?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/v1/payment?page=1"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "&laquo; Previous"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": false
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/v1/payment"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a new payment",
                "operationId": "a0f91d816de17822d493fecdb9b517c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_client",
                                    "cpf",
                                    "method_payment",
                                    "amount"
                                ],
                                "properties": {
                                    "name_client": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "minLength": 3,
                                        "example": "John Doe"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "maxLength": 11,
                                        "minLength": 11,
                                        "example": "12345678901"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 3,
                                        "example": "Payment for service"
                                    },
                                    "method_payment": {
                                        "type": "string",
                                        "enum": [
                                            "pix",
                                            "boleto",
                                            "bank_transfer"
                                        ],
                                        "example": "boleto"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 150.75
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "uuid",
                                                    "example": "01972371-f5ee-7262-bc61-ae757a7ceeb8"
                                                },
                                                "name_client": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "12345678901"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Payment for service"
                                                },
                                                "method_payment": {
                                                    "type": "string",
                                                    "example": "boleto"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150.75
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-30 22:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}